Rest/requirements.txt
appdirs==1.4.0
Django==1.10.5
djangorestframework==3.5.3
MySQL-python==1.2.5
packaging==16.8
PyJWT==1.4.2
pyparsing==2.1.10
six==1.10.0
